@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_platform_channels/data/native/channel.dart';
 
@@ -9,12 +8,13 @@ abstract class NativeDataSource {
 class NativeDataSourceImpl extends NativeDataSource {
   final MethodChannel methodChannel;
 
-  NativeDataSourceImpl({@required this.methodChannel});
+  NativeDataSourceImpl({required this.methodChannel});
 
   @override
   Future<int> getTemperature() async {
     try {
-      final int currentTemp = await methodChannel.invokeMethod(Channel.getTemp);
+      final int? currentTemp =
+          await methodChannel.invokeMethod(Channel.getTemp);
       if (currentTemp == null) {
         throw PlatformException(message: 'temp is missing', code: "");
       }
